Fixed reward share percentage bug.
When creating a 5% reward share transaction, the UI was calling the core's API with a value of `500000000`, but it should have instead used `5_00`. This caused the transaction to be rejected due to being far above the maximum limit of 100_00 (100%). The same problem applied when removing a reward share transaction. Thanks to yopilax for discovering this bug.