modify electron dependencies to static versions
This fixes the recent issues with regard to running the UI with electron. Some dependencies went up in versions and something broke, this will force the dependencies to use versions that worked until we come up with a more 'in-depth' solution for the long term.

Also updated the git repo URL to correct new repo location.