Increased polling intervals in minting screen from 5s to 30s
These were adding quite a lot of load to the core and causing high CPU usage. The downside of slowing these requests down is it could exaggerate the incorrect stats problem caused by incorrect (low) info being returned from the core in relation to the number of minters per level. This will need to be fixed in the core very soon.