Publishing a website now requires a zip to be supplied by the user.
This is a workaround of the previous limitation requiring the UI and core to be on the same machine. Now that a zip file is required rather than a local path, it can be passed to the core regardless of the machine it is running on. Longer term we can handle the zipping or serialization of multiple files automatically in the UI, but not in the first version.