Added "Hosted Content" screen, for an easy way to review and delete hosted content.
When content is from a name that is followed, an "UNFOLLOW" button is shown in place of "DELETE". This is because there is no point in deleting followed content as it will be re-fetched right away. Once unfollowed, the delete button shows. This might be a bit confusing in terms of UX, but it's better than nothing for a first version.

Note: relies on new core APIs.